[general] This commit mostly fixes issues noticed/caused by vs2012 changed BitmapRegion() to check if GlobalAlloc() returns NULL before using the results. changed CreateGrayscaleIcon() to check if the icons icInfo.hbmColor is NULL before trying to use it. changed CreateGrayscaleIcon() to check if icInfo.hbmMask is NULL before deleting it. changed TGetWindowText() to use nText+1 when allocating string, instead of ++nText, avoids potential overrun. added new regex option to isRegexMatch() that uses C++11's regex objects. changed dll.def to use Version()'s decorated name due to a conflict in vs2012 added Windows 7 + IE 10 + GDI+ 1.1 to OS Versions list in defines.h bumped dev build to 20 changed PFNISTHEMEACTIVE to return a BOOL
[mIRCLinker]
changed initMapFile() to check m_hFileMap is non NULL before using it.

[xpopup]
changed XPopupMenuManager::parseIdentifier() to use a token list & switch() combo instead of all the if()'s
added unknown prop error message to
XPopupMenuManager::parseIdentifier()

[xtreebar]
changed $xtreebar(mIRC).item to make sure szbuf is zero terminated.

[tstring]
changed constructors to set m_savedtotaltoks to zero.
changed some vars to consts
changed deltok()/puttok() to check if p_cEnd == NULL first, this fixes possible use of NULL pointer.
changed WcharTochar() to use l+1 when allocating the result, fixes possible buffer overrun issue.

[tinyxml]
changed TiXmlNode destructor to call Clear() function.
changed TiXmlNode::Clear() function to use a for loop.

[scroll]
changed $xdid().line to only use a single %d in the wnsprinf()'s pattern, to match the single arg given.

[box]
changed WM_VSCROLL/WM_HSCROLL handling to check for lParam == NULL first.

[webctrl]
changed constructors fail handling to correctly check for pointers being NULL before trying to Release() them.
added some debug code.

[treeview]
changed /xdid -w to check for himl being NULL before using it.
changed /xdid -w to check for icon being NULL before destroying it.
changed the structure of DCXTVSORT to include the text buffers used by the sort function, this avoids these large-ish buffers being on the stack.
changed /xdid -z to allocate DCXTVSORT on the heap as its now too big for the stack.
changed insertItem() to set tvi.cchTextMax to the length of itemtext instead of the odd sizeof() sum it was (why was it this sum?....)
changed sortItemsEx() to use the text buffers in ptvsort.

[statusbar][rebar]
changed /xdid -w to check for himl being NULL before using it.

[dcxml]
changed parseDialog() to check for parent being NULL before using & if it is NULL set parenttype as NULL too.
add idElement == NULL check to parseId()

[listview]
changed the structure of DCXLVSORT to include the text buffers used by the sort function, this avoids these large-ish buffers being on the stack.
changed /xdid -z to allocate DCXLVSORT on the heap as its now too big for the stack.
changed sortItemsEx() to use the text buffers in plvsort.
changed $xdid().gtext to make sure wstr is zero terminated.
changed matchItemText() to make sure itemtext is zero terminated.
changed NM_CLICK handling to include subitem select code in all builds not just debug.

[dialog]
changed WM_DRAWITEM handling to check for idata == NULL first.
changed SC_SIZE handling to zero ret var first.
changed UpdateShadow() to check if hbitmap or hMemDC == NULL before using.

[controls]
changed resizeBitmap() to check if the arg srcBM == NULL & return NULL if it does.

[comboex]
changed CBN_SELENDOK handling to make sure itemtext is zero terminated.

[divider]
changed WM_CREATE handling to only SetProp() AFTER lpdvdata has been setup.

[xtray]
changed some vars to be const's
changed TrayIcon() to use XSwitchFlags.

git-svn-id: https://dcx.googlecode.com/svn/branches/dcxutf@1115 17a06da3-8c20-fa79-550f-e9588e724b77

+++ Classes/UltraDock/dcxDockCommands.cpp
@@ -605,17 +605,17 @@ mIRC(_xdock)
 			break;
 		case 12: // switchBarHwnd
 			{
-				wnsprintf(data, MIRC_BUFFER_SIZE_CCH, TEXT("%d"), Dcx::mIRC.getSwitchbar());
+				wnsprintf(data, MIRC_BUFFER_SIZE_CCH, TEXT("%ld"), (DWORD)Dcx::mIRC.getSwitchbar()); // don't use %p as this gives a hex result.
 			}
 			break;
 		case 13: // toolBarHwnd
 			{
-				wnsprintf(data, MIRC_BUFFER_SIZE_CCH, TEXT("%d"), Dcx::mIRC.getToolbar());
+				wnsprintf(data, MIRC_BUFFER_SIZE_CCH, TEXT("%ld"), (DWORD)Dcx::mIRC.getToolbar());
 			}
 			break;
 		case 14: // treeBarHwnd
 			{
-				wnsprintf(data, MIRC_BUFFER_SIZE_CCH, TEXT("%d"), Dcx::mIRC.getTreebar());
+				wnsprintf(data, MIRC_BUFFER_SIZE_CCH, TEXT("%ld"), (DWORD)Dcx::mIRC.getTreebar());
 			}
 			break;
 		case 0: // error
